refactor(protocol): move note ID helpers from standards
