+++ CHANGELOG.md
@@ -13,8 +13,8 @@
 ### Changes
 
 - [BREAKING] Reduced the maximum number of assets a note can carry from 64 to 16 ([#3381](https://github.com/0xMiden/protocol/issues/3381)).
+- Moved the note ID helpers into the protocol library as `miden::protocol::input_note::find_note` and `miden::protocol::output_note::compute_note_id`. The new `INPUT_NOTE_INDEX_REQUEST_EVENT` lets transaction hosts provide an input-note index hint: successful lookups authenticate it against the `NoteId` cached by the transaction prologue, while reported misses are validated by a full scan. This makes the common found case O(1) in the VM while preserving soundness, and provides a shared fast lookup for `FeeSponsorshipNote` and the fee-collection work tracked in [#3340](https://github.com/0xMiden/protocol/issues/3340) ([#3424](https://github.com/0xMiden/protocol/pull/3424)).
 - [BREAKING] Transaction summaries now bind the reference block, expiration delta, and seven user parameters; the Rust and MASM APIs changed accordingly ([#3210](https://github.com/0xMiden/protocol/issues/3210)).
-- Optimized `miden::standards::note::note_id::find_input_note_by_id` with the new `INPUT_NOTE_INDEX_REQUEST_EVENT`: executor and prover hosts now provide a candidate input-note index, successful lookups are authenticated against the `NoteId` cached by the transaction prologue, and reported misses are validated by a full scan. This makes the common found case O(1) in the VM while preserving soundness, and provides a shared fast lookup for `FeeSponsorshipNote` and the fee-collection work tracked in [#3340](https://github.com/0xMiden/protocol/issues/3340) ([#3424](https://github.com/0xMiden/protocol/pull/3424)).
 - Moved account-patch commitment validation from `AccountUpdateDetails::validate()` into `TxAccountUpdate::new()` and consolidated all `ProvenTransaction` invariant checks in `from_parts()`, fixing a deserialization bypass of the circular-note check ([#3412](https://github.com/0xMiden/protocol/pull/3412)).
 - [BREAKING] Renamed `TransactionContext` to `MockTransaction` and `TxContextInput` to `MockTransactionInput` in `miden-testing`, and migrated the transaction tests to `MockChain::build_transaction` ([#3313](https://github.com/0xMiden/protocol/pull/3313)).
 - [BREAKING] Removed `AccountBuilder::with_auth_component`; the authentication component is now passed like any other component via `with_component` or `with_components` ([#3379](https://github.com/0xMiden/protocol/pull/3379)).
@@ -42,7 +42,7 @@
 - [BREAKING] Made the AggLayer bridge's network ID a deployment setting: the ID is stored in the `agglayer::bridge::network_id` storage slot (written once at account creation and read at runtime via `bridge_config::load_network_id`) instead of being compiled in as the `MIDEN_NETWORK_ID` MASM constant, so all deployments share one bridge code commitment; `create_bridge_account` now takes a `network_id` argument ([#3062](https://github.com/0xMiden/protocol/pull/3062)).
 - Added the `tx::get_fee_faucet_id` kernel accessor, exposing the fee faucet ID from the transaction's reference block to user code ([#2899](https://github.com/0xMiden/protocol/discussions/2899)).
 - Added the `miden::protocol::tx::compute_fee` procedure, which lets account and note code compute the transaction fee during execution ([#3211](https://github.com/0xMiden/protocol/issues/3211)).
-- Added the `FeeSponsorshipNote` standard note, which carries the fee for the feature note it is bound to by `NoteId`, and a `miden::standards::note::note_id` MASM module for computing note IDs on-chain ([#3274](https://github.com/0xMiden/protocol/pull/3274)).
+- Added the `FeeSponsorshipNote` standard note, which carries the fee for the feature note it is bound to by `NoteId`, along with helpers for computing note IDs on-chain ([#3274](https://github.com/0xMiden/protocol/pull/3274)).
 - Added the `UpgradeManager` account component for network account code and storage upgrades ([#3299](https://github.com/0xMiden/protocol/pull/3299)).
 - Added the `miden::standards::assets::non_fungible_asset::validate` MASM procedure, which validates a non-fungible asset's composition and the binding of its value to the asset class, and used it in the `NonFungibleFaucet` burn procedure ([#3308](https://github.com/0xMiden/protocol/pull/3308)).
 - Added a `FeeManager` account component exposing the FPI-callable `estimate_note_fee` procedure, dispatching the fee computation to a configurable fee policy (first policy: `ConstantFeePolicy`) switchable via the authority-gated `set_fee_policy` ([#3309](https://github.com/0xMiden/protocol/pull/3309), [#3328](https://github.com/0xMiden/protocol/issues/3328)).
