+++ crates/miden-protocol/asm/protocol/src/input_note.masm
@@ -1,3 +1,5 @@
+use miden::core::word
+use miden::protocol::tx
 use {INPUT_NOTE_GET_ASSET_OFFSET, INPUT_NOTE_GET_ATTACHMENTS_COMMITMENT_OFFSET, INPUT_NOTE_GET_INITIAL_ASSETS_INFO_OFFSET, INPUT_NOTE_GET_METADATA_OFFSET, INPUT_NOTE_GET_NOTE_ID_OFFSET, INPUT_NOTE_GET_RECIPIENT_OFFSET, INPUT_NOTE_GET_SCRIPT_ROOT_OFFSET, INPUT_NOTE_GET_SERIAL_NUMBER_OFFSET, INPUT_NOTE_GET_STORAGE_INFO_OFFSET, INPUT_NOTE_REMOVE_ALL_ASSETS_OFFSET, INPUT_NOTE_REMOVE_ASSET_OFFSET}
     from miden::protocol::kernel_proc_offsets
 use miden::protocol::note
@@ -12,6 +14,17 @@ use {AccountId, Bool, MemoryAddress, NoteId, NoteMetadata, NoteRecipient, NoteSc
 #! The host places [note_idx, is_found] on the advice stack. Because this response is
 #! unauthenticated, callers must validate both reported matches and misses.
 pub const INPUT_NOTE_INDEX_REQUEST_EVENT = event("miden::protocol::input_note::index_request")
+
+# CONSTANTS
+# =================================================================================================
+
+#! Local memory offset at which `find_note` stages the note ID it searches for.
+const TARGET_NOTE_ID_LOC = 0
+
+# ERRORS
+# =================================================================================================
+
+const ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID = "input note index lookup result from the host is invalid"
 
 # PROCEDURES
 # =================================================================================================
@@ -496,6 +509,86 @@ pub proc get_note_id_raw(is_active_note: Bool, note_index: u16) -> NoteId
     # clean the stack
     swapdw dropw dropw swapw dropw
     # => [NOTE_ID]
+end
+
+#! Finds the input note with the specified ID among the transaction's input notes.
+#!
+#! The host first supplies an index and a presence flag. A reported match is validated in O(1) by
+#! comparing the note ID cached by the prologue at the supplied index. If the host reports no
+#! match, every input note is inspected to prove that the note is absent.
+#!
+#! Inputs:  [NOTE_ID]
+#! Outputs: [is_found, note_idx]
+#!
+#! Where:
+#! - NOTE_ID is the ID of the input note to look for.
+#! - is_found is 1 if an input note with the specified ID exists, 0 otherwise.
+#! - note_idx is the index of the matching input note; it is only meaningful if is_found is 1.
+#!
+#! Panics if:
+#! - the host reports a matching index whose cached note ID differs from NOTE_ID.
+#! - the host reports that NOTE_ID is absent when it is present among the input notes.
+#!
+#! Invocation: exec
+@locals(4)
+pub proc find_note(note_id: NoteId) -> (Bool, u16)
+    loc_storew_le.TARGET_NOTE_ID_LOC
+    # => [NOTE_ID]
+
+    emit.INPUT_NOTE_INDEX_REQUEST_EVENT dropw
+    # => []
+
+    adv_push adv_push
+    # => [is_found, note_idx]
+
+    if.true
+        # Authenticate a reported match against the note ID cached by the prologue.
+        dup exec.get_note_id
+        # => [NOTE_ID, note_idx]
+
+        padw loc_loadw_le.TARGET_NOTE_ID_LOC
+        # => [TARGET_NOTE_ID, NOTE_ID, note_idx]
+
+        exec.word::eq assert.err=ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID
+        # => [note_idx]
+
+        push.1
+        # => [is_found = 1, note_idx]
+    else
+        # The host-reported index is meaningless when no match was reported.
+        drop
+        # => []
+
+        exec.tx::get_num_input_notes
+        # => [num_notes]
+
+        dup neq.0
+        # => [should_loop, num_notes]
+
+        while.true
+            # The counter holds the number of notes not yet inspected; the note at counter - 1 is
+            # inspected next, so after sub.1 the value doubles as that note's index.
+            sub.1
+            # => [note_idx]
+
+            dup exec.get_note_id
+            # => [NOTE_ID, note_idx]
+
+            padw loc_loadw_le.TARGET_NOTE_ID_LOC
+            # => [TARGET_NOTE_ID, NOTE_ID, note_idx]
+
+            exec.word::eq not assert.err=ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID
+            # => [note_idx]
+
+            dup neq.0
+            # => [should_loop, note_idx]
+        end
+        # => [note_idx]
+
+        drop push.0 push.0
+        # => [is_found = 0, note_idx = 0]
+    end
+    # => [is_found, note_idx]
 end
 
 #! Returns the metadata of the input note with the specified index.
