+++ crates/miden-standards/asm/standards/note/mod.masm
@@ -13,7 +13,6 @@ use {NoteRecipient, NoteScriptRoot} from miden::protocol::types
 
 pub mod execution_hint
 pub mod note_creator
-pub mod note_id
 
 # ERRORS
 # =================================================================================================
