--- crates/miden-standards/asm/standards/note/note_id.masm
@@ -1,177 +1,0 @@
-#! miden::standards::note::note_id
-#!
-#! Computes output note IDs and finds input notes by ID.
-#!
-#! Input and active note IDs are exposed directly by `miden::protocol::input_note::get_note_id` and
-#! `miden::protocol::active_note::get_note_id`. Output notes have no such accessor, so their ID is
-#! computed from the four commitments the kernel exposes, mirroring the kernel's own derivation.
-
-use miden::core::crypto::hashes::poseidon2
-use miden::core::word
-
-use miden::protocol::input_note
-use miden::protocol::output_note
-use miden::protocol::tx
-use {INPUT_NOTE_INDEX_REQUEST_EVENT} from miden::protocol::input_note
-
-# CONSTANTS
-# ================================================================================================
-
-#! Local memory offset at which find_input_note_by_id stages the note ID it searches for.
-const TARGET_NOTE_ID_LOC = 0
-
-# ERRORS
-# ================================================================================================
-
-const ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID = "input note index lookup result from the host is invalid"
-
-# PROCEDURES
-# ================================================================================================
-
-#! Computes the note ID from the four note commitments.
-#!
-#! The note ID is computed as:
-#! - NOTE_METADATA_COMMITMENT = hash(METADATA || ATTACHMENTS_COMMITMENT)
-#! - NOTE_DETAILS_COMMITMENT  = hash(RECIPIENT || ASSETS_COMMITMENT)
-#! - NOTE_ID                  = hash(NOTE_DETAILS_COMMITMENT || NOTE_METADATA_COMMITMENT)
-#!
-#! Inputs:  [METADATA, ATTACHMENTS_COMMITMENT, RECIPIENT, ASSETS_COMMITMENT]
-#! Outputs: [NOTE_ID]
-#!
-#! Where:
-#! - METADATA is the note's metadata word.
-#! - ATTACHMENTS_COMMITMENT is the commitment to the note's attachments.
-#! - RECIPIENT is the commitment to the note's serial number, script root and storage.
-#! - ASSETS_COMMITMENT is the commitment to the note's assets.
-#! - NOTE_ID is the resulting note ID.
-#!
-#! Invocation: exec
-proc compute_note_id
-    exec.poseidon2::merge
-    # => [NOTE_METADATA_COMMITMENT, RECIPIENT, ASSETS_COMMITMENT]
-
-    movdnw.2
-    # => [RECIPIENT, ASSETS_COMMITMENT, NOTE_METADATA_COMMITMENT]
-
-    exec.poseidon2::merge
-    # => [NOTE_DETAILS_COMMITMENT, NOTE_METADATA_COMMITMENT]
-
-    exec.poseidon2::merge
-    # => [NOTE_ID]
-end
-
-#! Computes the ID of the output note at the provided index.
-#!
-#! The returned ID is only final once the note has been fully constructed, that is, once all of its
-#! assets and attachments have been added.
-#!
-#! Inputs:  [note_idx]
-#! Outputs: [NOTE_ID]
-#!
-#! Where:
-#! - note_idx is the index of the output note whose ID should be computed.
-#! - NOTE_ID is the ID of the specified output note.
-#!
-#! Panics if:
-#! - the note index is greater or equal to the total number of output notes.
-#!
-#! Invocation: exec
-pub proc compute_output_note_id
-    dup exec.output_note::get_assets_info
-    # => [ASSETS_COMMITMENT, num_assets, note_idx]
-
-    movup.4 drop movup.4
-    # => [note_idx, ASSETS_COMMITMENT]
-
-    dup exec.output_note::get_recipient movup.4
-    # => [note_idx, RECIPIENT, ASSETS_COMMITMENT]
-
-    dup exec.output_note::get_attachments_commitment movup.4
-    # => [note_idx, ATTACHMENTS_COMMITMENT, RECIPIENT, ASSETS_COMMITMENT]
-
-    exec.output_note::get_metadata
-    # => [METADATA, ATTACHMENTS_COMMITMENT, RECIPIENT, ASSETS_COMMITMENT]
-
-    exec.compute_note_id
-    # => [NOTE_ID]
-end
-
-#! Finds the input note with the specified ID among the transaction's input notes.
-#!
-#! The host first supplies an index and a presence flag. A reported match is validated in O(1) by
-#! comparing the note ID cached by the prologue at the supplied index. If the host reports no
-#! match, every input note is inspected to prove that the note is absent.
-#!
-#! Inputs:  [NOTE_ID]
-#! Outputs: [is_found, note_idx]
-#!
-#! Where:
-#! - NOTE_ID is the ID of the input note to look for.
-#! - is_found is 1 if an input note with the specified ID exists, 0 otherwise.
-#! - note_idx is the index of the matching input note; it is only meaningful if is_found is 1.
-#!
-#! Panics if:
-#! - the host reports a matching index whose cached note ID differs from NOTE_ID.
-#! - the host reports that NOTE_ID is absent when it is present among the input notes.
-#!
-#! Invocation: exec
-@locals(4)
-pub proc find_input_note_by_id
-    loc_storew_le.TARGET_NOTE_ID_LOC
-    # => [NOTE_ID]
-
-    emit.INPUT_NOTE_INDEX_REQUEST_EVENT dropw
-    # => []
-
-    adv_push adv_push
-    # => [is_found, note_idx]
-
-    if.true
-        # Authenticate a reported match against the note ID cached by the prologue.
-        dup exec.input_note::get_note_id
-        # => [NOTE_ID, note_idx]
-
-        padw loc_loadw_le.TARGET_NOTE_ID_LOC
-        # => [TARGET_NOTE_ID, NOTE_ID, note_idx]
-
-        exec.word::eq assert.err=ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID
-        # => [note_idx]
-
-        push.1
-        # => [is_found = 1, note_idx]
-    else
-        # The host-reported index is meaningless when no match was reported.
-        drop
-        # => []
-
-        exec.tx::get_num_input_notes
-        # => [num_notes]
-
-        dup neq.0
-        # => [should_loop, num_notes]
-
-        while.true
-            # The counter holds the number of notes not yet inspected; the note at counter - 1 is
-            # inspected next, so after sub.1 the value doubles as that note's index.
-            sub.1
-            # => [note_idx]
-
-            dup exec.input_note::get_note_id
-            # => [NOTE_ID, note_idx]
-
-            padw loc_loadw_le.TARGET_NOTE_ID_LOC
-            # => [TARGET_NOTE_ID, NOTE_ID, note_idx]
-
-            exec.word::eq not assert.err=ERR_INPUT_NOTE_INDEX_LOOKUP_INVALID
-            # => [note_idx]
-
-            dup neq.0
-            # => [should_loop, note_idx]
-        end
-        # => [note_idx]
-
-        drop push.0 push.0
-        # => [is_found = 0, note_idx = 0]
-    end
-    # => [is_found, note_idx]
-end
