+++ crates/miden-standards/asm/standards/notes/fee_sponsorship.masm
@@ -1,10 +1,10 @@
 use miden::protocol::account_id
 use miden::protocol::active_account
 use miden::protocol::active_note
+use miden::protocol::input_note
 use miden::protocol::note
 use miden::protocol::tx
 
-use miden::standards::note::note_id
 use miden::standards::wallets::basic as basic_wallet
 
 use {AccountId, BlockNumber, NoteRecipient, NoteId, NoteTag, NoteType} from miden::protocol::types
@@ -153,7 +153,7 @@ pub proc main
     # script rather than in the consuming account because the sponsor trusts neither the consuming
     # account nor the transaction builder. The lookup authenticates a host-reported match against
     # the prologue-cached note ID and scans all input notes to validate a reported miss.
-    exec.note_id::find_input_note_by_id
+    exec.input_note::find_note
     # => [is_found, feature_note_idx, pad(16)]
 
     swap drop
