+++ crates/miden-testing/src/standards/mod.rs
@@ -1,6 +1,5 @@
 mod account_upgrade;
 mod fee_policy_manager;
 mod network_account_target;
-mod note_id;
 mod note_tag;
 mod token_metadata;
